Move LiteralNoneViolation into a separate module.
