+++ tests/conftest.py
@@ -9,6 +9,7 @@ import pytest
 
 from wemake_python_styleguide.options.config import Configuration
 from wemake_python_styleguide.violations import (
+    annotations,
     best_practices,
     complexity,
     consistency,
@@ -47,6 +48,7 @@ def _load_all_violation_classes():
         best_practices,
         refactoring,
         oop,
+        annotations,
     ]
 
     classes = {}
