+++ tests/fixtures/noqa.py
@@ -562,8 +562,8 @@ assert []  # noqa: WPS444
 unhashable = [] * 2  # noqa: WPS435
 
 
-def literal_none_func(arg: Literal[None]):  # noqa: WPS347
+def literal_none_func(arg: Literal[None]):  # noqa: WPS701
     '''Literal[None]'''
 
-def literal_none_return_func() -> Literal[None]:  # noqa: WPS347
+def literal_none_return_func() -> Literal[None]:  # noqa: WPS701
     '''Literal[None]'''
